Refactor MultiJson to reduce size
