+++ lib/multi_json.rb
@@ -3,9 +3,11 @@ require_relative "multi_json/version"
 require_relative "multi_json/adapter_error"
 require_relative "multi_json/parse_error"
 require_relative "multi_json/options_cache"
+require_relative "multi_json/adapter_selector"
 
 module MultiJson
   include Options
+  extend AdapterSelector
   extend self
 
   def default_options=(value)
@@ -37,24 +39,9 @@ module MultiJson
     gson: "gson"
   }.freeze
 
-  # The default adapter based on what you currently
-  # have loaded and installed.
-  def default_adapter
-    adapter = loaded_adapter || installable_adapter
-    return adapter if adapter
-
-    @default_adapter_warning_shown ||= begin
-      Kernel.warn(
-        "[WARNING] MultiJson is using the default adapter (ok_json). " \
-        "We recommend loading a different JSON library to improve performance."
-      )
-      true
-    end
-
-    :ok_json
+  class << self
+    alias_method :default_engine, :default_adapter
   end
-
-  alias_method :default_engine, :default_adapter
 
   # Get the current adapter class.
   def adapter
@@ -66,16 +53,6 @@ module MultiJson
   end
   alias_method :engine, :adapter
 
-  # Set the JSON parser utilizing a symbol, string, or class.
-  # Supported by default are:
-  #
-  # * <tt>:fast_jsonparser</tt>
-  # * <tt>:oj</tt>
-  # * <tt>:json_gem</tt>
-  # * <tt>:ok_json</tt>
-  # * <tt>:yajl</tt>
-  # * <tt>:gson</tt> (JRuby only)
-  # * <tt>:jr_jackson</tt> (JRuby only)
   def use(new_adapter)
     @adapter = load_adapter(new_adapter)
   ensure
@@ -84,23 +61,7 @@ module MultiJson
   alias_method :adapter=, :use
   alias_method :engine=, :use
 
-  def load_adapter(new_adapter)
-    case new_adapter
-    when String, Symbol then load_adapter_from_string_name new_adapter.to_s
-    when NilClass, FalseClass then load_adapter default_adapter
-    when Class, Module then new_adapter
-    else raise ::LoadError, new_adapter
-    end
-  rescue ::LoadError => e
-    raise(AdapterError.build(e), cause: e)
-  end
 
-  # Decode a JSON string into Ruby.
-  #
-  # <b>Options</b>
-  #
-  # <tt>:symbolize_keys</tt> :: If true, will use symbols instead of strings for the keys.
-  # <tt>:adapter</tt> :: If set, the selected adapter will be used for this call.
   def load(string, options = {})
     adapter = current_adapter(options)
     begin
@@ -134,36 +95,4 @@ module MultiJson
     self.adapter = old_adapter
   end
   alias_method :with_engine, :with_adapter
-
-  private
-
-  # Checks for already loaded adapters and returns the first match
-  def loaded_adapter
-    return :fast_jsonparser if defined?(::FastJsonparser)
-    return :oj if defined?(::Oj)
-    return :yajl if defined?(::Yajl)
-    return :jr_jackson if defined?(::JrJackson)
-    return :json_gem if defined?(::JSON::Ext::Parser)
-    return :gson if defined?(::Gson)
-
-    nil
-  end
-
-  # Attempts to load and return the first installable adapter
-  def installable_adapter
-    REQUIREMENT_MAP.each do |adapter, library|
-      require library
-      return adapter
-    rescue ::LoadError
-      next
-    end
-    nil
-  end
-
-  def load_adapter_from_string_name(name)
-    normalized_name = ALIASES.fetch(name, name).to_s
-    require "multi_json/adapters/#{normalized_name.downcase}"
-    klass_name = normalized_name.split("_").map(&:capitalize).join
-    MultiJson::Adapters.const_get(klass_name)
-  end
 end
